Single photo
Added click listener to items in collection.
When clicked, a photo fragment is added on top of collection.
Using RxJava PublishSubject to register for clicks from
adapter back to collection fragment.

To make this possible, the main collection view is not
a fragment instead.

Also made some name changes in layouts and layout id's.