Animations
* Added explode content animation when leaving collection.
* Added fade content animation when leaving photo fragment.
* Added date information to items in collection whenever
  a new day is found.
* Quite a lot of renaming as well.

Placed under debug flag a way to make the first photo in
a day group to span while screen. Works in some sence
but there is some strange things going on with
the recyclerview, staggered layout manager and when
adapter is asked to bind view. When binding the view
the loading of the bitmap is started, but by checking
logs, there is a view measuremnt made after picasso
have set the bitmap. So this is not activated but kept
behind flags.