DiskPicasso loading
Updated disk picasso to be able to fetch all cache entries
for a given file path.

The cache entries are memory cached for instant retrievals.

Updated recyclerview adapter to use this and load with picasso
with resize instead. This is done by knowing either the
horizontal or vertical bound for the layout together with
view span to pre calculate the required width and height
of the bitmap.

Btw, there seems to be a bug in picasso when dealing with
images that have orientation in exif, the fit and centerFit
scaled the image wrongly, so this is a neccessary change
from that point of view.