Fragment listeners
Added fragment listeners back to MainActivity.

Since the RecyclerViewFragment is resused, the best way
was to make it abstract and subclass it with a

  CarouselFragment
  MainGridFragment

and let those have their own click listeners instead. This way
I could keep the way google suggests that fragments talk with
MainActivity (through a class cast in onAttach).

With this refactor, I could also remove span, orientation that
was set through bundles to the fragment(s) and use abstract
methods instead. Much nices to fragment lifecycle.