Disk cache key
Previously the image source path was used to get entries
from the cache. However, if the image was replaces by another
image with the same source path, the cache would return an
"old" cached entry.

Fixed this by require a source file key. In Android, the
_ID in MediaStore is ever increasing and I basically construct
the source file key from the source file path appended with the
_ID to create a file key.

Also fixed a bug when the cache is evicting which lead to that
after the storage and journal was cleared, there was a callback
to the memory cache again.